ce_cloud_models/PaddleHub/NLP/linux/requirements.txt
sentencepiece==0.1.95
regex
tqdm==4.57.0
jieba
seqeval
attrdict
h5py
paddlenlp
colorama
colorlog
easydict
filelock
flask >= 1.1.0
gitpython
numpy
matplotlib
opencv-python
packaging
paddle2onnx >= 0.5.1
Pillow
pyyaml
pyzmq
rarfile
tqdm
visualdl >= 2.0.0
# gunicorn not support windows
gunicorn >= 19.10.0; sys_platform != "win32"
